python_interface/requirements.txt
pyserial
comtypes
pycaw